Add support for R9M stock bootloader (#178)
* Add support for R9M stock bootloader

This adds a secondary bootloader on top of the stock one. This 2nd BL uses STK500 as it's upload method and requires an OpenTX version that supports it (see https://github.com/opentx/opentx/pull/7968).

* Reduced timeout to 300ms and exit STK500 on missing initial STK_GET_SYNC

This basically make the bootloader exit very fast most of the time (CRSF triggers random bytes received). In case nothing in received, exit after 300ms.

* copy firmware.bin to firmware.elrs so it can be flashed by OpenTx

* fixed HSE_VALUE

* Use 400kbps for debug no matter which platform